Refactor: namespace 정리
