+++ nodeedge/model/__init__.py
@@ -1,1 +1,2 @@
-from ._model import AbstractModel, Model
+from ._model import Model, LinkPropertyModel
+from ._base_model import AbstractModel
